backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# OCR and ML dependencies
pandas>=1.3.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
pillow-heif>=0.10.0
python-doctr>=0.6.0
sentence-transformers>=2.2.0
torch>=1.12.0
torchvision>=0.13.0

# Google services
gspread>=5.7.0
oauth2client>=4.1.3
PyDrive>=1.3.1

# Utilities
python-dateutil>=2.8.0
PyYAML>=6.0.1
